+++ src/actions/publishAndSync.js
@@ -1,13 +1,6 @@
 import { useDocumentOperation, useSchema, useClient } from 'sanity'
-import { buildSyncPayload } from '../utils/sync'
-import { querySiblings } from '../utils/sibling'
-import { syncToSiblings } from '../utils/sync'
 import apiVersion from '../apiVersion'
 
-/**
- * Document action component for "Publish & Sync"
- * Syncs shared fields to all sibling translations, then publishes the document
- */
 export function publishAndSyncAction(props) {
   const { type, id, draft } = props
   const operation = useDocumentOperation(id, type)
@@ -15,7 +8,6 @@ export function publishAndSyncAction(props) {
   const client = useClient({apiVersion})
   const schemaType = schema.get(type)
   const syncFieldNames = schemaType?.options?.kaliber?.syncFieldNames ?? []
-
   
   if (syncFieldNames.length === 0)
     return null
@@ -28,6 +20,7 @@ export function publishAndSyncAction(props) {
 
         if (siblings.length > 0) {
           const syncPayload = buildSyncPayload(draft, syncFieldNames)
+
           await syncToSiblings(siblings, syncPayload, client)
         }
 
@@ -38,4 +31,70 @@ export function publishAndSyncAction(props) {
     },
     disabled: !draft
   }
+}
+
+/**
+ * Query all sibling translations of a document
+ * @param {string} documentType - Document type
+ * @param {string} translationId - Translation ID to find siblings for
+ * @param {string} currentLanguage - Current document's language
+ * @param {any} sanityClient - Sanity client
+ * @returns {Promise<Array<any>>} Array of sibling documents
+ */
+export async function querySiblings(documentType, translationId, currentLanguage, sanityClient) {
+  try {
+    if (!translationId) {
+      return []
+    }
+
+    const siblings = await sanityClient.fetch(
+      `*[
+        _type == $type &&
+        translationId == $translationId &&
+        language != $currentLanguage &&
+        !(_id in path("drafts.**"))
+      ]`,
+      {
+        type: documentType,
+        translationId,
+        currentLanguage
+      }
+    )
+
+    return siblings || []
+  } catch (error) {
+    throw new Error(`Failed to query sibling translations: ${error.message}`)
+  }
+}
+
+/**
+ * Build sync payload from fields
+ * @param {any} document - Current document state
+ * @param {Array<string>} syncFieldNames - Field names marked for syncing
+ * @returns {Object} Payload with field values
+ */
+export function buildSyncPayload(document, syncFieldNames) {
+  return Object.fromEntries(
+    syncFieldNames.map(fieldName => [fieldName, document[fieldName]])
+  )
+}
+
+/**
+ * Patch all siblings with sync payload in a single transaction
+ * @param {Array<any>} siblings - Array of sibling documents
+ * @param {Object} payload - Payload to patch
+ * @param {any} sanityClient - Sanity client
+ * @returns {Promise<void>}
+ */
+export async function syncToSiblings(siblings, payload, sanityClient) {
+  if (siblings.length === 0)
+    return
+
+  const transaction = sanityClient.transaction()
+
+  siblings.forEach(sibling => {
+    transaction.patch(sibling._id, patch => patch.set(payload))
+  })
+
+  await transaction.commit()
 }
